Devices/PC/DevQemuFwCfg.cpp: Improve efficiency of the RAM based framebuffer refresh callback by mapping the framebuffer once and copying directly from the memory buffer instead of going through PDMDevHlpPhysReadUser() which requires PGM locking, etc., github:gh-34 [build fix]
svn:sync-xref-src-repo-rev: r170298

+++ src/VBox/Devices/PC/DevQemuFwCfg.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevQemuFwCfg.cpp 110778 2025-08-21 10:19:05Z alexander.eichner@oracle.com $ */
+/* $Id: DevQemuFwCfg.cpp 110779 2025-08-21 10:23:14Z alexander.eichner@oracle.com $ */
 /** @file
  * DevQemuFwCfg - QEMU firmware configuration compatible device.
  */
@@ -1757,7 +1757,7 @@ static DECLCALLBACK(int) qemuFwCfgR3RamfbPortUpdateDisplay(PPDMIDISPLAYPORT pInt
             if (RT_UNLIKELY(!pThis->Ramfb.cPgLocks))
             {
                 size_t const cbFb = pThis->RamfbCfg.cbStride * pThis->RamfbCfg.cHeight;
-                uint32_t const cPages = (cbFb + GUEST_PAGE_SIZE - 1) >> GUEST_PAGE_SHIFT;
+                uint32_t const cPages = (uint32_t)((cbFb + GUEST_PAGE_SIZE - 1) >> GUEST_PAGE_SHIFT);
                 RTGCPHYS GCPhysStart = pThis->RamfbCfg.GCPhysRamfbBase & ~(RTGCPHYS)GUEST_PAGE_OFFSET_MASK;
                 RTGCPHYS const offPage = pThis->RamfbCfg.GCPhysRamfbBase & (RTGCPHYS)GUEST_PAGE_OFFSET_MASK;
